Data-Extraction/requirements.txt
fastmcp>=0.1.0
python-dotenv>=1.0.0
openai>=1.10.0
python-doctr[torch]>=0.8.0
docling>=1.0.0
pydantic>=2.0.0
uvicorn>=0.20.0
rapidocr_onnxruntime
